# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Breathe ESG data ingestion and review platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{id}/uploads:
    get:
      operationId: listCompanyUploads
      tags: [uploads]
      summary: List uploads for a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Uploads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataUpload"
  /companies/{id}/uploads/{sourceType}:
    post:
      operationId: uploadData
      tags: [uploads]
      summary: Upload data file for a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sourceType
          in: path
          required: true
          schema:
            type: string
            enum: [sap, utility, travel]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "201":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid data
        "404":
          description: Company not found
components:
  schemas:
...
    DataUpload:
      type: object
      required: [id, companyId, sourceType, filename, status, totalRows, importedRows, errorRows, uploadedAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        sourceType:
          type: string
        filename:
          type: string
        status:
          type: string
          enum: [processing, completed, failed]
        totalRows:
          type: integer
        importedRows:
          type: integer
        errorRows:
          type: integer
        uploadedAt:
          type: string
          format: date-time
        errorLog:
          type: ["string", "null"]
    UploadInput:
      type: object
      required: [filename, csvData]
      properties:
        filename:
          type: string
        csvData:
          type: string
          description: Raw CSV content as string
    UploadResult:
      type: object
      required: [upload, importedRows, errorRows, errors]
      properties:
        upload:
          $ref: "#/components/schemas/DataUpload"
        importedRows:
          type: integer
        errorRows:
          type: integer
        errors:
          type: array
          items:
            type: string